Move factor_rotation unit test to /tests folder
