+++ statsmodels/multivariate/factor_rotation/_analytic_rotation.py
@@ -152,124 +152,3 @@ def promax(A,k=2):
     T=np.linalg.inv(S.dot(D)).T
     #return
     return A.dot(T), T
-
-class unittests(unittest.TestCase):
-    
-    @staticmethod    
-    def str2matrix(A):
-        A=A.lstrip().rstrip().split('\n')
-        A=np.array([row.split() for row in A]).astype(np.float)
-        return A
-        
-    def test_target_rotation(self):
-        """
-        Rotation towards target matrix example
-        http://www.stat.ucla.edu/research/gpa
-        """
-        A= self.str2matrix("""
-         .830 -.396
-         .818 -.469
-         .777 -.470
-         .798 -.401
-         .786  .500
-         .672  .458
-         .594  .444
-         .647  .333
-        """)
-        H= self.str2matrix("""
-          .8 -.3
-          .8 -.4
-          .7 -.4
-          .9 -.4
-          .8  .5
-          .6  .4
-          .5  .4
-          .6  .3
-        """)
-        T = target_rotation(A,H)
-        L = A.dot(T)
-        L_required = self.str2matrix("""
-        0.84168  -0.37053
-        0.83191  -0.44386
-        0.79096  -0.44611
-        0.80985  -0.37650
-        0.77040   0.52371
-        0.65774   0.47826
-        0.58020   0.46189
-        0.63656   0.35255
-        """)
-        self.assertTrue(np.allclose(L,L_required,atol=1e-05))
-        T = target_rotation(A,H,full_rank=True)
-        L = A.dot(T)
-        self.assertTrue(np.allclose(L,L_required,atol=1e-05))
-
-    def test_orthogonal_target(self):
-        """
-        Rotation towards target matrix example
-        http://www.stat.ucla.edu/research/gpa
-        """
-        import _gpa_rotation as gr
-        A= self.str2matrix("""
-         .830 -.396
-         .818 -.469
-         .777 -.470
-         .798 -.401
-         .786  .500
-         .672  .458
-         .594  .444
-         .647  .333
-        """)
-        H= self.str2matrix("""
-          .8 -.3
-          .8 -.4
-          .7 -.4
-          .9 -.4
-          .8  .5
-          .6  .4
-          .5  .4
-          .6  .3
-        """)
-        vgQ = lambda L=None, A=None, T=None: gr.vgQ_target(H,L=L,A=A,T=T)
-        L, phi, T, table = gr.GPA(A, vgQ=vgQ, rotation_method='orthogonal')
-        T_analytic = target_rotation(A,H)
-        self.assertTrue(np.allclose(T,T_analytic,atol=1e-05))
-
-
-if __name__ == '__main__': # run only if this file is run directly and not when imported
-    run_unit_tests=True
-    test_only = list() # if list is empty then test all
-    #test_only.append('test_orthogonal_target')
-    if run_unit_tests:
-        if len(test_only) > 0:
-            suite = unittest.TestSuite()
-            for ut in test_only:
-                suite.addTest(unittests(ut))
-            unittest.TextTestRunner().run(suite)
-        else:
-            #unittest.main()
-            suite = unittest.TestLoader().loadTestsFromTestCase(unittests)
-            unittest.TextTestRunner(verbosity=2).run(suite)
-    else: # run a basic example
-        A= unittests.str2matrix("""
-         .830 -.396
-         .818 -.469
-         .777 -.470
-         .798 -.401
-         .786  .500
-         .672  .458
-         .594  .444
-         .647  .333
-        """)
-        H= unittests.str2matrix("""
-          .8 -.3
-          .8 -.4
-          .7 -.4
-          .9 -.4
-          .8  .5
-          .6  .4
-          .5  .4
-          .6  .3
-        """)
-        T = target_rotation(A,H)
-        print(T)
-        L, T = promax(A)
