+++ statsmodels/multivariate/factor_rotation/_wrappers.py
@@ -310,99 +310,3 @@ def rotate_factors(A, method, *method_args, **algorithm_kwargs):
         L=rotateA(A,T,rotation_method=rotation_method)
     #return
     return L, T
-
-class unittests(unittest.TestCase):
-    
-    @staticmethod    
-    def str2matrix(A):
-        A=A.lstrip().rstrip().split('\n')
-        A=np.array([row.split() for row in A]).astype(np.float)
-        return A
-        
-    def get_A(self):
-        return self.str2matrix("""
-         .830 -.396
-         .818 -.469
-         .777 -.470
-         .798 -.401
-         .786  .500
-         .672  .458
-         .594  .444
-         .647  .333
-        """)
-        
-    def get_H(self):
-        return  self.str2matrix("""
-          .8 -.3
-          .8 -.4
-          .7 -.4
-          .9 -.4
-          .8  .5
-          .6  .4
-          .5  .4
-          .6  .3
-        """)
-
-    def get_W(self):
-        return self.str2matrix("""
-        1 0
-        0 1
-        0 0
-        1 1
-        1 0
-        1 0
-        0 1
-        1 0
-        """)
-
-        
-    def _test_template(self, method,*method_args, **algorithms):
-        A=self.get_A()
-        algorithm1= 'gpa' if 'algorithm1' not in algorithms else algorithms['algorithm1']
-        algorithm2= 'gpa_der_free' if 'algorithm`' not in algorithms else algorithms['algorithm1']
-        L1, T1 = rotate_factors(A,method,*method_args, algorithm = algorithm1)        
-        L2, T2 = rotate_factors(A,method,*method_args, algorithm = algorithm2)
-        self.assertTrue(np.allclose(L1, L2, atol=1e-5))
-        self.assertTrue(np.allclose(T1, T2, atol=1e-5))
-    
-    def test_methods(self):
-        """
-        Quartimax derivative free example
-        http://www.stat.ucla.edu/research/gpa
-        """
-        #orthomax, oblimin and CF are tested indirectly
-        methods=['quartimin', 'biquartimin',
-                 'quartimax', 'biquartimax', 'varimax', 'equamax', 'parsimax', 'parsimony',
-                 'target', 'partial_target']
-        for method in methods:
-            method_args=[]
-            if method == 'target':
-                method_args=[self.get_H(),'orthogonal']
-                self._test_template(method, *method_args)
-                method_args=[self.get_H(),'oblique']
-                self._test_template(method, *method_args)
-                method_args=[self.get_H(),'orthogonal']
-                self._test_template(method, *method_args, algorithm2='analytic')
-            elif method == 'partial_target':
-                method_args=[self.get_H(), self.get_W()]
-            self._test_template(method, *method_args)
-    
-if __name__ == '__main__': # run only if this file is run directly and not when imported
-    run_unit_tests=True
-    test_only = list() # if list is empty then test all
-    #test_only.append('test_orthogonal_target')
-    if run_unit_tests:
-        if len(test_only) > 0:
-            suite = unittest.TestSuite()
-            for ut in test_only:
-                suite.addTest(unittests(ut))
-            unittest.TextTestRunner().run(suite)
-        else:
-            #unittest.main()
-            suite = unittest.TestLoader().loadTestsFromTestCase(unittests)
-            unittest.TextTestRunner(verbosity=2).run(suite)
-    else: # run a basic example
-        A = np.random.randn(8,2)
-        L, T = rotate_factors(A,'varimax')
-        L, T = rotate_factors(A,'orthomax',0.5)
-        L, T = rotate_factors(A,'quartimin')
